requirements.txt
aiohappyeyeballs          2.4.3
aiohttp                   3.11.8
aiosignal                 1.3.1
altair                    5.5.0
annotated-types           0.7.0
anyio                     4.6.2.post1
asttokens                 3.0.0
attrs                     24.2.0
beautifulsoup4            4.12.3
blinker                   1.9.0
cachetools                5.5.0
certifi                   2024.8.30
charset-normalizer        3.4.0
click                     8.1.7
colorama                  0.4.6
contourpy                 1.3.1
coqpit                    0.0.17
cycler                    0.12.1
dataclasses-json          0.6.7
decorator                 5.1.1
deta                      1.2.0
distro                    1.9.0
docopt                    0.6.2
docx2txt                  0.8
executing                 2.1.0
faiss-cpu                 1.9.0.post1
fastapi                   0.115.6
filelock                  3.16.1
fonttools                 4.55.0
frozenlist                1.5.0
fsspec                    2024.10.0
gitdb                     4.0.11
GitPython                 3.1.43
gTTS                      2.5.4
h11                       0.14.0
httpcore                  1.0.7
httpx                     0.28.0
httpx-sse                 0.4.0
huggingface-hub           0.26.3
idna                      3.10
1.4.7
langchain                 0.3.9
langchain-community       0.3.8
langchain-core            0.3.21
langchain-text-splitters  0.3.2
langsmith                 0.1.147
markdown-it-py            3.0.0
MarkupSafe                3.0.2
marshmallow               3.23.1
matplotlib                3.9.3
matplotlib-inline         0.1.7
mdurl                     0.1.2
mpmath                    1.3.0
multidict                 6.1.0
mypy-extensions           1.0.0
narwhals                  1.14.3
networkx                  3.4.2
numpy                     1.26.4
openai                    0.28.0
orjson                    3.10.12
packaging                 24.2
pandas                    2.2.3
parso                     0.8.4
pillow                    11.0.0
pip                       24.2
pipwin                    0.5.2
prompt_toolkit            3.0.48
propcache                 0.2.0
pyarrow                   18.1.0
PyAudio                   0.2.14
pydantic                  2.10.2
pydantic_core             2.27.1
pydantic-settings         2.6.1
pydeck                    0.9.1
pydub                     0.25.1
Pygments                  2.18.0
pyjsparser                2.7.1
pyparsing                 3.2.0
PyPDF2                    3.0.1
PyPrind                   2.11.3
pySmartDL                 1.3.4
python-dateutil           2.9.0.post0
python-dotenv             1.0.1
python-multipart          0.0.19
pytz                      2024.2
PyYAML                    6.0.2
referencing               0.35.1
regex                     2024.11.6
requests                  2.32.3
requests-toolbelt         1.0.0
rich                      13.9.4
rpds-py                   0.21.0
safetensors               0.4.5
sentencepiece             0.2.0
setuptools                75.6.0
six                       1.16.0
smmap                     5.0.1
sniffio                   1.3.1
soupsieve                 2.6
SQLAlchemy                2.0.35
stack-data                0.6.3
starlette                 0.41.3
streamlit                 1.40.2
sympy                     1.13.1
tenacity                  9.0.0
tiktoken                  0.8.0
tokenizers                0.20.3
toml                      0.10.2
torch                     2.6.0.dev20241128+cpu
tornado                   6.4.2
tqdm                      4.67.1
traitlets                 5.14.3
transformers              4.46.3
typing_extensions         4.12.2
typing-inspect            0.9.0
tzdata                    2024.2
tzlocal                   5.2
urllib3                   2.2.3
uvicorn                   0.32.1
watchdog                  6.0.0
wcwidth                   0.2.13
wheel                     0.45.1
yarl                      1.18.0